Feedback Part 2 - Move BrowserSessionsObserver
